// src/shared/api/models/getApiInsightsUniqueCustomers200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiInsightsUniqueCustomers200DataItem } from './getApiInsightsUniqueCustomers200DataItem';

export type GetApiInsightsUniqueCustomers200 = {
    success: boolean;
    data: GetApiInsightsUniqueCustomers200DataItem[];
    message: string;
    statusCode: number;
};