// src/shared/api/models/getApiInsightsUniqueCustomers200DataItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiInsightsUniqueCustomers200DataItem = {
    employeeId: number;
    employeeName: string;
    totalUniqueCustomers: number;
    totalRevenue: number;
    averageRevenuePerCustomer: number;
    totalVisits: number;
    averageVisitsPerCustomer: number;
    newCustomersCount: number;
    returningCustomersCount: number;
};